+++ cmd/gophermart/users/users.go
@@ -20,8 +20,8 @@ type User struct {
 	ID        PostgresPK `json:"user_id,omitempty" db:"user_id"`
 	Login     string     `json:"login,omitempty" db:"login"`
 	Password  string     `json:"password,omitempty" db:"password"`
-	Current   int16      `json:"current" db:"current"`
-	Withdrawn int16      `json:"withdrawn" db:"withdrawn"`
+	Current   float32    `json:"current" db:"current"`
+	Withdrawn float32    `json:"withdrawn" db:"withdrawn"`
 }
 
 const (
@@ -34,7 +34,7 @@ const (
 type UserOrder struct {
 	UserID     PostgresPK `json:"user_id,omitempty" db:"user_id"`
 	Number     string     `json:"number" db:"number"`
-	Accrual    int16      `json:"accrual,omitempty" db:"accrual"`
+	Accrual    float32    `json:"accrual,omitempty" db:"accrual"`
 	Status     string     `json:"status" db:"status"`
 	UploadedAt time.Time  `json:"uploaded_at" db:"uploaded_at"`
 }
@@ -42,6 +42,6 @@ type UserOrder struct {
 type UserWithdraw struct {
 	UserID      PostgresPK `json:"user_id,omitempty" db:"user_id"`
 	Number      string     `json:"order" db:"number"`
-	Sum         int16      `json:"sum,omitempty" db:"sum"`
+	Sum         float32    `json:"sum,omitempty" db:"sum"`
 	ProcessedAt time.Time  `json:"processed_at" db:"processed_at"`
 }
